Better organize pollers and mocks into modules
